Commit non-working types with table
