+++ src/pages/admin/website/data.tsx
@@ -1,11 +1,15 @@
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/router'
 import type { FC } from 'react'
-import { useEffect, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import type { Dataset } from '@prisma/client'
+import type { Company } from '@prisma/client'
+import { useQuery, useQueryClient } from '@tanstack/react-query'
+import { getQueryKey } from '@trpc/react-query'
 
 import { AdminNavBar } from '../../../components'
 import { DynamicPaginatedAdminTable } from '../../../components/Table/DynamicPaginatedAdminTable'
+import { INDUSTRIES } from '../../../utils/constants'
 import { api } from '../../../utils/api'
 
 const UpdateData: FC = () => {
@@ -13,17 +17,44 @@ const UpdateData: FC = () => {
   const { push } = useRouter()
   const [dataset, setDataSet] = useState<Dataset>('COMPANY')
   const [columns, setColumns] = useState([])
-  const [rows, setRows] = useState([])
-  const { data, isLoading } = api.company.getCompaniesBySkipTake.useQuery({
-    skip: 0,
+  const [rows, setRows] = useState<Company[]>([])
+  const [skip, setSkip] = useState(0)
+
+  const { data: investment } = api.investment.getInvestmentsBySkipTake.useQuery(
+    {
+      skip: skip,
+      take: 20,
+    },
+  )
+  const { data: company } = api.company.getCompaniesBySkipTake.useQuery({
+    skip: skip,
     take: 20,
   })
+
+  console.log(investment)
+  const industryKeyValue = INDUSTRIES.map((industry) => ({
+    key: industry,
+    value: industry,
+  }))
 
   useEffect(() => {
     if (status === 'unauthenticated') {
       void push('/auth/error')
     }
   }, [push, status])
+
+  const originalRows = useMemo(() => company?.items ?? [], [company?.items])
+
+  // // Add data to table when data is loaded
+  useEffect(() => {
+    const cleanedRows = originalRows.map((row) => {
+      return {
+        ...row,
+        changedEntries: [],
+      }
+    })
+    setRows(cleanedRows)
+  }, [originalRows, setRows])
 
   return (
     session && (
@@ -32,9 +63,76 @@ const UpdateData: FC = () => {
         <button onClick={() => setDataSet('INVESTMENT')}>Investments</button>
         <button onClick={() => setDataSet('COMPANY')}>Companies</button>
         <DynamicPaginatedAdminTable
+          skip={skip}
+          setSkip={setSkip}
           dataset={dataset}
-          originalRows={[]}
-          columns={[]}
+          originalRows={rows}
+          columns={[
+            {
+              key: 'name',
+              label: 'Name',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.name ?? '',
+              },
+            },
+            {
+              key: 'ticker',
+              label: 'Ticker',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.ticker ?? '',
+              },
+            },
+            {
+              key: 'sector',
+              label: 'Sector',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.sector ?? '',
+              },
+            },
+            {
+              key: 'industry',
+              label: 'Industry',
+              isEditable: true,
+              ctrl: {
+                type: 'select',
+                options: industryKeyValue,
+                render: (row) => row.industry ?? '',
+              },
+            },
+            {
+              key: 'description',
+              label: 'Description',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.description ?? '',
+              },
+            },
+            {
+              key: 'netAssetVal',
+              label: 'Net Asset Value',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.netAssetVal,
+              },
+            },
+            {
+              key: 'bloombergId',
+              label: 'Bloomberg ID',
+              isEditable: true,
+              ctrl: {
+                type: 'text',
+                render: (row) => row.bloombergId ?? '',
+              },
+            },
+          ]}
         />
       </div>
     )
